BUG: Prevents os.remove from masking an error with open in get_contents_to_filename
When exceptions were raised because open failed, the exception handler
would call os.remove(filename) which would raise an exception saying
that the file does not exist. This would mask the exception that open
raised.

By checking for the existence of the file on disk prior to removing it,
these types of exceptions should no longer be masked.

+++ boto/s3/key.py
@@ -1706,7 +1706,9 @@ class Key(object):
                                           res_download_handler=res_download_handler,
                                           response_headers=response_headers)
         except Exception:
-            os.remove(filename)
+            if os.path.exists(filename):
+                # If open fails, then filename doesn't exist.
+                os.remove(filename)
             raise
         # if last_modified date was sent from s3, try to set file's timestamp
         if self.last_modified is not None:
